backennd/requirements.txt
Flask==.3.0.0
biopython==1.82
requests==2.31.0
ete3==3.1.3
numpy==1.26.3
six==1.16.0
termcolor==2.4.0
matplotlib==3.8.3
